refactor(ngwaf/scope): pull scope helpers into their own file
